refactor
move initialization of calculator list to SuntimesCalculatorDescriptor; keep only weakReference to context
